Feat : 판매자 이름 수정 기능 구현, (구매자/판매자) 이름 수정 메소드 통합
