+++ src/main/java/com/mypill/domain/buyer/controller/BuyerController.java
@@ -5,9 +5,7 @@ import com.mypill.domain.address.entity.Address;
 import com.mypill.domain.address.service.AddressService;
 import com.mypill.domain.comment.entity.Comment;
 import com.mypill.domain.comment.service.CommentService;
-import com.mypill.domain.member.service.MemberService;
 import com.mypill.domain.order.dto.response.OrderListResponse;
-import com.mypill.domain.order.dto.response.OrderResponse;
 import com.mypill.domain.order.entity.Order;
 import com.mypill.domain.order.service.OrderService;
 import com.mypill.domain.post.entity.Post;
@@ -18,9 +16,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.List;
 
@@ -28,7 +24,6 @@ import java.util.List;
 @RequiredArgsConstructor
 @RequestMapping("/usr/buyer")
 public class BuyerController {
-    private final MemberService memberService;
     private final PostService postService;
     private final CommentService commentService;
     private final OrderService orderService;
@@ -97,12 +92,5 @@ public class BuyerController {
         model.addAttribute("addresses",addressResponses);
 
         return "usr/buyer/myAddress";
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @ResponseBody
-    @PostMapping("/name/update")
-    public String nameUpdate(String newName) {
-        return memberService.nameUpdate(rq.getMember(), newName);
     }
 }
