+++ src/main/java/com/mypill/domain/member/controller/MemberController.java
@@ -12,7 +12,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
@@ -81,5 +80,11 @@ public class MemberController {
         SecurityContextHolder.clearContext();
         request.getSession().invalidate();
         response.sendRedirect("/");
+    }
+    @PreAuthorize("isAuthenticated()")
+    @ResponseBody
+    @PostMapping("/name/update")
+    public String nameUpdate(String newName) {
+        return memberService.nameUpdate(rq.getMember(), newName);
     }
 }
