+++ src/main/resources/templates/usr/buyer/myInfo.html
@@ -39,7 +39,7 @@
                 return;
             }
 
-            const url = '/usr/consumer/name/update';
+            const url = '/usr/member/name/update';
             const csrfToken = $('meta[name=csrf-token]').attr('content'); // 메타 태그에서 CSRF 토큰 가져오기
 
             // Ajax 요청을 사용하여 댓글 수정 API 호출
